Transform server values to not-include null
- Even when `null` are removed, the cache is updated as expected
- Send `null` for values that are to be cleared on forms
