+++ src/utils/common.ts
@@ -23,3 +23,22 @@ export function isValidUrl(url: string | undefined): url is string {
     const sanitizedUrl = url.replace(rege, 'localhost.com');
     return isValidRemoteUrl(sanitizedUrl);
 }
+
+/*
+type Clean<T> = (
+    T extends (infer Z)[]
+        ? Clean<Z>[]
+        : (
+            T extends object
+                ? { [K in keyof T]: Clean<T[K]> }
+                : (T extends null ? undefined : T)
+        )
+)
+
+type testType = {
+    numbers?: (null | number)[] | null; name: number | null;
+    age: number;
+    meta: { username: string | null } | null;
+}
+type test = Clean<testType>
+*/
