+++ src/utils/errorTransform.ts
@@ -66,22 +66,3 @@ function transformArray(errors: ArrayError[] | undefined): BaseError | undefined
 export const transformToFormError = transformObject;
 
 // const errors: ObjectError[] = <get_from_server>;
-
-/*
-type Clean<T> = (
-    T extends (infer Z)[]
-        ? Clean<Z>[]
-        : (
-            T extends object
-                ? { [K in keyof T]: Clean<T[K]> }
-                : (T extends null ? undefined : T)
-        )
-)
-
-type testType = {
-    numbers?: (null | number)[] | null; name: number | null;
-    age: number;
-    meta: { username: string | null } | null;
-}
-type test = Clean<testType>
-*/
